prepare moving graph.ui to graph extension
